added two new images, tweaked with spacing and decorated a little more
two new images replaced the black bars on the sides labeled "filler". these images are not fully styled and are temporary regarding their placement, however they will be two of the images in use.
made the list items more spaced out for improved legibility.
added a horizontal rule within the header for decoration and seperation between title and content. also acts as a ruler for which images should not pass above.